requirements.txt
# requirements.txt
prefect>=2.0
python-dotenv==1.0.0